Add EMBEDDED_CLI_SERIAL_XLATE option (#20)
* Add EMBEDDED_CLI_SERIAL_XLATE option

When this option is set then CR will be considered (along with NL)
as ending a command, and NLs will be translated to CR NLs.
This makes things work seamlessly with serial terminals.

* Update posix_demo.c to use raw mode

This updates posix_demo.c to use raw mode (which is typical
of what's done when using serial ports). In particular it will
cause CR to be sent when ENTER is pressed.